Add files via upload
Empty webpage with Javascript Array function with comments inside the code. To see the code read the page in notepad or use developer tools of the particular web browser, To see the results of the code you have to open that webpage than click right click of touch pad or mouse and click inspect. Than go to second window and click console. You should have a list of pet names printed there. Code is explained in comments text between these // brackets in a bit better details to understand it better.